src/requirements.txt
editdistance
matplotlib
natsort
numpy
opencv-python
Pillow
python-Levenshtein
python-telegram-bot
scikit-learn
tensorboard